feat postgresql: fixed uneven load distribution on kRoundRobin policy
With the postgres kRoundRobin host selection policy, the selection is carried out using an atomic counter and the remainder of the division by the number of hosts.

This approach was supposed to ensure uniform distribution, but the problem is that the array of hosts is sorted by RTT once a second.

Because of this, such an algorithm could create a skew in the load on certain database hosts.

Tests: idk
commit_hash:fb6929f67524d84d6865fd5425859214f28f23ae